requirements.txt
azure-functions
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
